requirements.txt
asgiref==3.10.0
attrs==25.3.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
PyJWT==2.10.1
PySocks==1.7.1
pytz==2025.2
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.5
selenium==4.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.11.0
wsproto==1.2.0
